requirements.txt
tensorflow-gpu==1.5.0
soundfile
librosa
pyworld
